+++ Direct3DWrapper/DX_NeuralNetwork/DxPooling.cpp
@@ -37,9 +37,9 @@ DxPooling::DxPooling(UINT width, UINT height, UINT poolNum, UINT inputsetnum) {
 	dx = Dx12Process::GetInstance();
 	mCommandList = dx->dx_sub[0].mCommandList.Get();
 	mObjectCB = new ConstantBuffer<CONSTANT_BUFFER_Pooling>(1);
-	cb.WidHei.x = Width;
-	cb.WidHei.y = Height;
-	cb.WidHei.z = PoolNum;
+	cb.WidHei.x = (float)Width;
+	cb.WidHei.y = (float)Height;
+	cb.WidHei.z = (float)PoolNum;
 	mObjectCB->CopyData(0, cb);
 }
